delete upd, get quiz  working

+++ src/main/java/com/sigma/service/impl/QuizServiceImpl.java
@@ -2,8 +2,11 @@ package com.sigma.service.impl;
 
 import com.sigma.model.dto.QuizDto;
 import com.sigma.model.entity.Quiz;
+import com.sigma.model.entity.Role;
+import com.sigma.model.entity.User;
 import com.sigma.repository.QuizRepository;
 import com.sigma.service.QuizService;
+import com.sigma.service.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -18,6 +21,8 @@ import java.util.List;
 public class QuizServiceImpl implements QuizService {
 
     private final QuizRepository quizRepository;
+
+    private final UserService userService;
 
     @Override
     public List<QuizDto> getAllQuizzes() {
@@ -38,32 +43,66 @@ public class QuizServiceImpl implements QuizService {
 
     @Override
     @Transactional
-    public Quiz createQuiz(QuizDto quiz) {
+    public Quiz createQuiz(QuizDto quiz, Long userId) {
+        User user = userService.findUserById(userId);
+        if (user == null) {
+            throw new EntityNotFoundException("User with id = " + userId + " not found");
+        }
+
+        if (user.getRole().equals(Role.CAPTAIN)) {
+            throw new EntityNotFoundException("This user is captain, not admin");
+        }
+
         log.info("Creating new quiz {}", quiz.toString());
         return quizRepository.save(QuizDto.toQuiz(quiz));
     }
 
     @Override
     @Transactional
-    public void updateQuiz(QuizDto updatedQuiz, Long quizId) {
+    public Quiz updateQuiz(QuizDto updatedQuiz, Long quizId, Long userId) {
+        User user = userService.findUserById(userId);
+        if (user == null) {
+            throw new EntityNotFoundException("User with id = " + userId + " not found");
+        }
+
+        if (user.getRole().equals(Role.CAPTAIN)) {
+            throw new EntityNotFoundException("This user is captain, not admin");
+        }
+
         QuizDto oldQuiz = QuizDto.fromQuiz(quizRepository.findById(quizId).get());
         if (oldQuiz == null) {
             throw new EntityNotFoundException();
         }
         log.info("Updating quiz {}", oldQuiz);
-        oldQuiz.setQuizName(updatedQuiz.getQuizName());
-        oldQuiz.setCategory(updatedQuiz.getCategory());
-        oldQuiz.setDateTime(updatedQuiz.getDateTime());
-        oldQuiz.setShortDescription(updatedQuiz.getShortDescription());
-
-        quizRepository.save(QuizDto.toQuiz(oldQuiz));
+        if (updatedQuiz.getQuizName() != null) {
+            oldQuiz.setQuizName(updatedQuiz.getQuizName());
+        }
+        if (updatedQuiz.getCategory() != null) {
+            oldQuiz.setCategory(updatedQuiz.getCategory());
+        }
+        if (updatedQuiz.getDateTime() != null) {
+            oldQuiz.setDateTime(updatedQuiz.getDateTime());
+        }
+        if (updatedQuiz.getShortDescription() != null) {
+            oldQuiz.setShortDescription(updatedQuiz.getShortDescription());
+        }
+        return quizRepository.save(QuizDto.toQuiz(oldQuiz));
     }
 
     @Override
     @Transactional
-    public void deleteQuiz(Long quizId) {
+    public void deleteQuiz(Long userId, Long quizId) {
+        User user = userService.findUserById(userId);
+        if (user == null) {
+            throw new EntityNotFoundException("User with id = " + userId + " not found");
+        }
+
+        if (user.getRole().equals(Role.CAPTAIN)) {
+            throw new EntityNotFoundException("This user is captain, not admin");
+        }
+
         if (!quizRepository.existsById(quizId)) {
-            throw new EntityNotFoundException();
+            throw new EntityNotFoundException("No quiz with this id");
         }
         log.info("Deleting quiz {}", quizRepository.getById(quizId));
         quizRepository.deleteById(quizId);
